[skip ci] Add tools script to list `torch` APIs
This adds a script that will run through everything in `dir(torch)` recursively and print out each path. We can run this for 1.8.1 and then again for 1.9.0 to determine the API difference between the two in order to ensure everything is properly documented and there are no secret BC-breaking changes.

ghstack-source-id: 9b26b49f80edae4f3e8d9b6b04e96b7636e4dc7b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/58610